@@ -1,8 +1,9 @@
 {
   "name": "gawk",
-  "version": "5.0.0",
+  "version": "6.0.0",
   "description": "Observable JavaScript object model",
-  "main": "./dist/index.js",
+  "type": "module",
+  "exports": "./src/index.js",
   "author": "Chris Barber <chris@cb1inc.com> (https://github.com/cb1kenobi)",
   "license": "MIT",
   "keywords": [
@@ -16,52 +17,30 @@
     "model"
   ],
   "scripts": {
-    "build": "gulp build",
-    "coverage": "gulp coverage",
-    "docs": "gulp docs",
-    "prepare": "gulp build",
-    "test": "gulp test"
+    "coverage": "c8 npm run test",
+    "lint": "eslint src test",
+    "prepack": "npm run lint",
+    "test": "npm run lint && mocha test/**/test-*.js --require test/setup.js --reporter spec"
   },
   "dependencies": {
-    "fast-deep-equal": "^3.1.3",
-    "source-map-support": "^0.5.19"
+    "fast-deep-equal": "^3.1.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.3",
-    "@babel/plugin-transform-modules-commonjs": "^7.12.1",
-    "@babel/register": "^7.12.1",
-    "ansi-colors": "^4.1.1",
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.2.1",
-    "babel-plugin-istanbul": "^6.0.0",
-    "chai": "^4.2.0",
-    "coveralls": "^3.1.0",
-    "esdoc": "^1.1.0",
-    "esdoc-ecmascript-proposal-plugin": "^1.0.0",
-    "esdoc-standard-plugin": "^1.0.0",
-    "eslint": "^7.13.0",
+    "c8": "^7.6.0",
+    "chai": "^4.3.0",
+    "eslint": "^7.20.0",
     "eslint-plugin-chai-expect": "^2.2.0",
     "eslint-plugin-mocha": "^8.0.0",
-    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-promise": "^4.3.1",
     "eslint-plugin-security": "^1.4.0",
-    "fancy-log": "^1.3.3",
-    "fs-extra": "^9.0.1",
-    "gulp": "^4.0.2",
-    "gulp-babel": "^8.0.0",
-    "gulp-debug": "^4.0.0",
-    "gulp-eslint": "^6.0.0",
-    "gulp-load-plugins": "^2.0.5",
-    "gulp-plumber": "^1.2.1",
-    "gulp-sourcemaps": "^3.0.0",
-    "mocha": "^8.2.1",
-    "nyc": "^15.1.0",
-    "sinon": "^9.2.1",
+    "mocha": "^8.3.0",
+    "sinon": "^9.2.4",
     "sinon-chai": "^3.5.0"
   },
   "homepage": "https://github.com/cb1kenobi/gawk",
   "bugs": "https://github.com/cb1kenobi/gawk/issues",
   "repository": "https://github.com/cb1kenobi/gawk",
   "engines": {
-    "node": ">=10.13.0"
+    "node": ">=12"
   }
 }
